+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - correctly selecting the input history item
 - improved task memory management in renderer process
 - correctly restoring stashed changes on failures when switching to worktree
+- added support for Solidity, Crystal and Haskel languages in semantic search
 
 ## [0.71.0]